Add archival deprecation notice to Brigade resource articles
Although it largely makes sense to keep old blog posts available for
archival reasons, we have had experiences lately of Brigade leaders
discovering these old resources and being confused. As we continue to
evolve the guidance we offer to brigades, it's important that any
brigades finding these archival pages via Google will be tipped off that
the content may no longer be the most up-to-date content we've written
on that topic.

To address this, I did a google search for [allinurl: blog brigade
site:codeforamerica.org] (a first cut at identifying the most likely
pages that a brigade leader would discover) and then added a disclaimer
to those pages pointing toward the existing Brigade website, which we
are positioning as the first stop for brigade leaders for the latest
information and resources.